V.1.4

+++ Source/IDEA.h
@@ -10,12 +10,12 @@
  modify it under the terms of the GNU Lesser General Public
  License as published by the Free Software Foundation; either
  version 2.1 of the License, or (at your option) any later version.
- 
+
  IDEA-cipher  is distributed in the hope that it will be useful,
  but WITHOUT ANY WARRANTY; without even the implied warranty of
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  Lesser General Public License for more details.
- 
+
  You should have received a copy of the GNU Lesser General Public
  License along with IDEA-cipher ; if not, write to the Free Software
  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
@@ -30,7 +30,6 @@
 #include "Round.h"
 #include "Keycreate.h"
 
-//AAA this functions destroies the key, user must use a copy
 
 //single instance encrypt fun, return 0 in case of error
 int IDEA_single_encrypt(uint16_t *message,uint16_t *key);
@@ -42,10 +41,10 @@ int IDEA_single_decrypt(uint16_t *message, uint16_t *key);
 void IDEA_crypt(uint16_t *message, uint16_t *subkey);
 
 //multi instance encrypt fun to process more blocks with same key, return number of blocks processed successfully, NO CBC
-int IDEA_multi_encrypt(uint16_t *message, uint16_t *key, const int blocks);
+unsigned long IDEA_multi_encrypt(uint16_t *message, uint16_t *key, const unsigned long blocks);
 
 //multi instance decrypt fun to process more blocks with same key, return number of blocks processed successfully, NO CBC
-int IDEA_multi_decrypt(uint16_t *message, uint16_t *key, const int blocks);
+unsigned long IDEA_multi_decrypt(uint16_t *message, uint16_t *key, const unsigned long blocks);
 
 //details in IDEA.c file
 #endif /* defined(__IDEA__IDEA__) */
